+++ Graphics/drawGlyph.cpp
@@ -264,7 +264,7 @@ void drawContext::drawEllipse(double x, double y, double z, float v0[3], float v
     v0[0], v0[1], v0[2], .0f,
     v1[0], v1[1], v1[2], .0f,
     v0[1]*v1[2]-v0[2]*v1[1], v0[2]*v1[0]-v0[0]*v1[2], v0[0]*v1[1]-v0[1]*v1[0], .0f,
-    x, y, z, 1.f
+    (GLfloat)x, (GLfloat)y, (GLfloat)z, 1.f
   };
   glMultMatrixf(m);
   glCallList(_displayLists + 2);
@@ -281,7 +281,7 @@ void drawContext::drawEllipsoid(double x, double y, double z, float v0[3], float
     v0[0], v0[1], v0[2], .0f,
     v1[0], v1[1], v1[2], .0f,
     v2[0], v2[1], v2[2], .0f,
-    x, y, z, 1.f
+    (GLfloat)x, (GLfloat)y, (GLfloat)z, 1.f
   };
   glMultMatrixf(m);
   glCallList(_displayLists + 0);
